Docker_Image_App/ccmgmt_data/requirements.txt
Flask
Flask-SQLAlchemy
DateTime
SQLAlchemy
psycopg2-binary
